+++ src/CuiLib/Converters/ValueConverter.Implementation.cs
@@ -2,7 +2,7 @@ using System;
 using System.IO;
 using System.Text;
 
-namespace CuiLib.Options
+namespace CuiLib.Converters
 {
     public static partial class ValueConverter
     {
