+++ src/CuiLib/Options/SingleValueOption.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using CuiLib.Checkers;
+using CuiLib.Converters;
 
 namespace CuiLib.Options
 {
