+++ src/Test/CuiLib/Options/ParameterTest.cs
@@ -1,5 +1,6 @@
 ﻿using CuiLib;
 using CuiLib.Checkers;
+using CuiLib.Converters;
 using CuiLib.Options;
 using NUnit.Framework;
 using System;
